requirements.txt
aiohttp==3.9.1
imgkit==1.2.3
prettytable==3.9.0
python-telegram-bot==20.7
Requests==2.31.0
